requirements.txt
algokit-utils==1.1.1
beaker-pyteal==1.0.0
cffi==1.15.1
docstring-parser==0.14.1
msgpack==1.0.5
py-algorand-sdk==2.2.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
pyteal==0.23.0
semantic-version==2.10.0
